split commands
